+++ Changelog.md
@@ -6,6 +6,7 @@ Released on ?.
 
 ✨ New Features / Improvements
 
+- Android: `geo:` links can now be opened with the app and adds a new location
 - Minimum Android version is now Android 12
 
 🐛 Bugfixes